Move mock data injection to its own module
